Automatically backed up by Learn

+++ lib/create.sql
@@ -7,7 +7,7 @@ CREATE TABLE users (
 CREATE TABLE projects (
   id INTEGER PRIMARY KEY,
   title TEXT,
-  funding_goal INTEGER,
+  funding_goal REAL,
   start_date TEXT,
   end_date TEXT,
 );
